Fundo not working

+++ Unity/Meros-Correnteza/Assets/Scripts/Objetos/MoveLeft.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class MoveLeft : MonoBehaviour
 {
-    [SerializeField] int speedX; // Velocidade
+    [SerializeField] float speedX; // Velocidade
     void Update()
     {
         transform.Translate(Vector2.left * Time.deltaTime * speedX); // Translata para a esquerda
